# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_photos_api_events__event_id__photos_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_upload_photos_api_events__event_id__photos_post",
        "type": "object"
      },
...
      "FeatureIn": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "FeatureIn",
        "type": "object"
      },
      "FeatureInfo": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "description",
          "enabled"
        ],
        "title": "FeatureInfo",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginIn": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
...
      "SheetPhoto": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "fx": {
            "title": "Fx",
            "type": "number"
          },
          "fy": {
            "title": "Fy",
            "type": "number"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "n_faces": {
            "title": "N Faces",
            "type": "integer"
          },
          "proc_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proc Ms"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "filename",
          "status",
          "n_faces",
          "proc_ms",
          "error",
          "width",
          "height",
          "fx",
          "fy"
        ],
        "title": "SheetPhoto",
        "type": "object"
      },
      "StatsOut": {
        "properties": {
          "avg_ms_per_photo": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Ms Per Photo"
          },
          "done": {
            "title": "Done",
            "type": "integer"
          },
          "errors": {
            "title": "Errors",
            "type": "integer"
          },
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "faces": {
            "title": "Faces",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "photos": {
            "title": "Photos",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "photos",
          "done",
          "pending",
          "errors",
          "faces",
          "avg_ms_per_photo"
        ],
        "title": "StatsOut",
        "type": "object"
      },
      "UploadResult": {
        "description": "Um item da resposta do upload: a foto salva (com `duplicate`) ou um erro\n({filename, status: \"error\", error}). A rota usa response_model_exclude_unset.",
        "properties": {
          "duplicate": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "n_faces": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Faces"
          },
          "proc_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proc Ms"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "status"
        ],
        "title": "UploadResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Foco",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/features/{key}": {
      "put": {
        "operationId": "set_feature_api_admin_features__key__put",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Feature",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/login": {
      "post": {
        "operationId": "login_api_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/events/{event_id}/photos": {
      "get": {
        "operationId": "list_photos_api_events__event_id__photos_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Event Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 2000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SheetPhoto"
                  },
                  "title": "Response List Photos Api Events  Event Id  Photos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Photos",
        "tags": [
          "photos"
        ]
      },
      "post": {
        "operationId": "upload_photos_api_events__event_id__photos_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_photos_api_events__event_id__photos_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UploadResult"
                  },
                  "title": "Response Upload Photos Api Events  Event Id  Photos Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Photos",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/events/{event_id}/progress": {
      "get": {
        "description": "Server-Sent Events: empurra o status das fotos até todas terminarem.\n\nSSE é só uma resposta HTTP que nunca fecha, onde o servidor escreve\nlinhas \"data: ...\\n\\n\". O navegador lê com `new EventSource(url)`.\nO status vive no Postgres (o worker é outro processo), então cada volta\nabre uma sessão curta e lê o banco.",
        "operationId": "progress_api_events__event_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Event Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "ids",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Progress",
        "tags": [
          "photos"
        ]
      }
    },
...
    "/api/photos/{photo_id}/full": {
      "get": {
        "operationId": "photo_full_api_photos__photo_id__full_get",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Photo Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "download",
            "required": false,
            "schema": {
              "default": false,
              "title": "Download",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Photo Full",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/photos/{photo_id}/medium": {
      "get": {
        "operationId": "photo_medium_api_photos__photo_id__medium_get",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Photo Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Photo Medium",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/photos/{photo_id}/thumb": {
      "get": {
        "operationId": "photo_thumb_api_photos__photo_id__thumb_get",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Photo Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Photo Thumb",
        "tags": [
          "photos"
        ]
      }
    },
...
    "/api/stats": {
      "get": {
        "operationId": "stats_api_stats_get",
        "parameters": [
          {
            "in": "query",
            "name": "event_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stats",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/zip": {
      "get": {
        "operationId": "download_zip_api_zip_get",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": true,
            "schema": {
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Zip",
        "tags": [
          "photos"
        ]
      }
    }
  }
}